contact_api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator
python-dotenv
fastapi-limiter
cloudinary
aioredis==2.0.0






